substrate refactor and restoring connection to substrate and centrifuge
